.ipynb_checkpoints/requirements-checkpoint.txt
streamlit==1.33.0
pandas
numpy
matplotlib
seaborn
pyyaml
scikit-learn
streamlit-authenticator==0.2.2